moving dummy data logic to new file
